+++ validator.go
@@ -15,12 +15,14 @@ type Validator struct {
 type ValidatorCollection []*Validator
 
 type IValidators interface {
-	getValidators(params ...interface{}) ValidatorCollection
+	GetValidators(params ...interface{}) ValidatorCollection
 }
 
 func Validate(obj IValidators, ctx ...interface{}) *ValidationError {
-	validationError := &ValidationError{}
-	validatorCollection := obj.getValidators(ctx...)
+	validationError := &ValidationError{
+		source: obj,
+	}
+	validatorCollection := obj.GetValidators(ctx...)
 	for _, validator := range validatorCollection {
 		if err := validator.Function(); err != nil {
 			validationError.Errors = append(validationError.Errors, &FieldError{
@@ -33,14 +35,13 @@ func Validate(obj IValidators, ctx ...interface{}) *ValidationError {
 	if len(validationError.Errors) > 0 {
 		return validationError
 	}
-
 	return nil
 }
 
 func GetValidatorsForList[T IValidators](fieldName string, list []T) ValidatorCollection {
 	validators := ValidatorCollection{}
 	for i, children := range list {
-		for _, validator := range children.getValidators() {
+		for _, validator := range children.GetValidators() {
 			childFieldName := strings.Join([]string{fieldName, fmt.Sprint(i), validator.Field}, ".")
 			newValidator := &Validator{Field: childFieldName, Function: validator.Function}
 			validators = append(validators, newValidator)
@@ -55,7 +56,7 @@ func GetValidatorsForObject[T IValidators](fieldName string, child T) ValidatorC
 	}
 
 	validators := ValidatorCollection{}
-	for _, validator := range child.getValidators() {
+	for _, validator := range child.GetValidators() {
 		childFieldName := strings.Join([]string{fieldName, validator.Field}, ".")
 		newValidator := &Validator{Field: childFieldName, Function: validator.Function}
 		validators = append(validators, newValidator)