+++ changelog.md
@@ -6,4 +6,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [v0.1.1] - 2022-11-19
+
+### Fixed
+- `Validate()` returned `ValidationError` without `source` field being set
+- `GetValidators()` was private method, therefore `Validate()` couldn't be used externally
+
+
 ## [v0.1.0] - 2022-11-19